+++ src/pydiffmap/utils.py
@@ -74,3 +74,38 @@ def sparse_from_fxn(X, K, function, Y=None):
 def _get_sparse_row_col(sparse_mat):
     sparse_mat = sparse_mat.tocoo()
     return sparse_mat.row, sparse_mat.col
+
+
+def _symmetrize_matrix(K, mode='average'):
+    """
+    Symmetrizes a sparse kernel matrix.
+
+    Parameters
+    ----------
+    K : scipy sparse matrix
+        The sparse matrix to be symmetrized, with positive elements on the nearest neighbors.
+    mode : string
+        The method of symmetrization to be implemented.  Current options are 'average', 'and', and 'or'.
+
+    Returns
+    -------
+    K_sym : scipy sparse matrix
+        Symmetrized kernel matrix.
+    """
+
+    if mode == 'average':
+        return 0.5*(K + K.transpose())
+    elif mode == 'or':
+        Ktrans = K.transpose()
+        dK = abs(K - Ktrans)
+        K = K + Ktrans
+        K = K + dK
+        return 0.5*K
+    elif mode == 'and':
+        Ktrans = K.transpose()
+        dK = abs(K - Ktrans)
+        K = K + Ktrans
+        K = K - dK
+        return 0.5*K
+    else:
+        raise ValueError('Did not understand symmetrization method')